+++ lib/commanded/commands/dispatcher.ex
@@ -5,7 +5,6 @@ defmodule Commanded.Commands.Dispatcher do
 
   alias Commanded.Aggregates.Aggregate
   alias Commanded.Aggregates.ExecutionContext
-  alias Commanded.Commands.ExecutionResult
   alias Commanded.Middleware.Pipeline
 
   defmodule Payload do
@@ -21,29 +20,28 @@ defmodule Commanded.Commands.Dispatcher do
       :handler_module,
       :handler_function,
       :aggregate_module,
-      :include_aggregate_version,
-      :include_execution_result,
       :identity,
       :identity_prefix,
       :timeout,
       :lifespan,
       :metadata,
       :retry_attempts,
+      :returning,
       middleware: []
     ]
   end
 
   @doc """
-  Dispatch the given command to the handler module for the aggregate as identified
-
-  Returns `:ok`, `{:ok, aggregate_version}` or 
-  `{:ok, %Commanded.Commands.ExecutionResult{}}` on success, or `{:error, error}`
-  on failure.
+  Dispatch the given command to the handler module for the aggregate as
+  identified.
   """
-  @spec dispatch(payload :: struct) :: :ok |
-    {:ok, Commanded.Commands.ExecutionResult.t()} |
-    {:ok, aggregate_version :: integer()} |
-    {:error, error :: term}
+  @spec dispatch(payload :: struct) ::
+          :ok
+          | {:ok, aggregate_state :: struct}
+          | {:ok, aggregate_version :: non_neg_integer()}
+          | {:ok, events :: list(struct)}
+          | {:ok, Commanded.Commands.ExecutionResult.t()}
+          | {:error, error :: term}
   def dispatch(%Payload{} = payload) do
     pipeline =
       payload
@@ -105,7 +103,14 @@ defmodule Commanded.Commands.Dispatcher do
         |> Pipeline.assign(:aggregate_version, aggregate_version)
         |> Pipeline.assign(:events, events)
         |> after_dispatch(payload)
-        |> respond_with_success(payload, events)
+        |> Pipeline.respond(:ok)
+
+      {:ok, aggregate_version, events, reply} ->
+        pipeline
+        |> Pipeline.assign(:aggregate_version, aggregate_version)
+        |> Pipeline.assign(:events, events)
+        |> after_dispatch(payload)
+        |> Pipeline.respond({:ok, reply})
 
       {:exit, {:normal, :aggregate_stopped}} ->
         # Maybe retry command when aggregate process stopped by lifespan timeout
@@ -138,7 +143,8 @@ defmodule Commanded.Commands.Dispatcher do
       handler_module: handler_module,
       handler_function: handler_function,
       lifespan: lifespan,
-      retry_attempts: retry_attempts
+      retry_attempts: retry_attempts,
+      returning: returning
     } = payload
 
     %ExecutionContext{
@@ -149,32 +155,9 @@ defmodule Commanded.Commands.Dispatcher do
       handler: handler_module,
       function: handler_function,
       lifespan: lifespan,
-      retry_attempts: retry_attempts
+      retry_attempts: retry_attempts,
+      returning: returning
     }
-  end
-
-  defp respond_with_success(%Pipeline{} = pipeline, payload, events) do
-    response =
-      case payload do
-        %{include_execution_result: true} ->
-          {
-            :ok,
-            %ExecutionResult{
-              aggregate_uuid: pipeline.assigns.aggregate_uuid,
-              aggregate_version: pipeline.assigns.aggregate_version,
-              events: events,
-              metadata: pipeline.metadata
-            }
-          }
-
-        %{include_aggregate_version: true} ->
-          {:ok, pipeline.assigns.aggregate_version}
-
-        _ ->
-          :ok
-      end
-
-    Pipeline.respond(pipeline, response)
   end
 
   defp before_dispatch(%Pipeline{} = pipeline, %Payload{middleware: middleware}) do